+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 - Allow mods that have no features to be displayed on the in-game options screen
 
+## 1.8.0
+
+- Split mixins into client and common subfolders for better filtering.
+- Wire up "client mode" to disable non-client mixins.
+
+## 1.7.5
+
+- Allow mods that have no features to be displayed on the in-game options screen
+
 ## 1.7.4
 
 - Disable dev environment check for debug.